README.md
Making up a game with dice of varying sides to practice basic JS functions. No classes or objects used, per instructions. 
I use some "brute force" loops where shorter code is possible because that's all I'm working on at this early point in my learning.

The rules  of the game are as follows:
A player (you) chooses to to play against 1,2, or 3 opponents. There is no way to play yourself, so you automatically play at least
one opponent and choose to add 0, 1, or 2 added opponents. Once the number of players is established, each player 6 dice simultaneously. 
The dice are of varying sides with one die each of 4, 6, 8, 10, 12, and 20 sides. If any of the dice match (i.e. the numbers produced by 
the function that rolls the dice are the same for any two dice), the player who rolled those dice is awarded points equal to the number of 
sides of the matching dice. For example, if you match the 4-sided and 10-sided dice, you get 4+10=14 points. The player with the largest
point total wins. Pretty simple. Eventually I may make this "prettier" and incorporate some innerHTML, but for now the game is displayed
in a browser's console.